将多个实体类中的金额字段类型统一更改为BigDecimal，以提高计算精度；更新相关逻辑以支持BigDecimal类型，优化CA余额、交易金额和费用计算的准确性，提升整体系统的用户体验。

+++ src/main/java/online/bingzi/aetherbot/events/ChatEventListener.java
@@ -142,7 +142,7 @@ public class ChatEventListener {
                 .divide(divisor, 9, RoundingMode.HALF_UP)
                 .multiply(multiplier);
                 
-            aiMessage.setCaCost(caCost.doubleValue());
+            aiMessage.setCaCost(caCost);
         }
         
         messageRepository.save(aiMessage);
@@ -176,7 +176,7 @@ public class ChatEventListener {
         // 记录CA交易
         CaTransaction transaction = new CaTransaction();
         transaction.setUser(user);
-        transaction.setAmount(actualCost.negate().doubleValue());
+        transaction.setAmount(actualCost.negate());
         transaction.setType(TransactionType.CONSUME);
         
         // 添加token使用量到交易描述中（如果有）
